--- components/Coach/Goal/Create/goalTemplates.tsx
@@ -1,229 +1,0 @@
-export const categories = [
-  "Education",
-  "Fitness",
-  "Career",
-  "Finance",
-  "Relationships",
-  "Hobbies",
-  "Mental health",
-];
-
-export const goals = [
-  {
-    name: "100 days of code",
-    stepName: "Code for 1 hour",
-    category: "Education",
-    description: "Code for 1 hour every day, for 100 days",
-    durationDays: 100,
-  },
-  {
-    name: "3 minutes of meditation",
-    stepName: "Meditate for 3 minutes",
-    category: "Mental health",
-    description: "Meditate for 3 minutes every day, for 10 days",
-    durationDays: 10,
-  },
-  {
-    name: "10 pushups every morning",
-    stepName: "Do 10 pushups",
-    category: "Fitness",
-    description: "Do 10 pushups every morning, for 30 days",
-    durationDays: 30,
-  },
-  {
-    name: "Read 1 chapter of a book",
-    stepName: "Read a chapter",
-    category: "Education",
-    description: "Read 1 chapter of a book every day, for 30 days",
-    durationDays: 30,
-  },
-  {
-    name: "Drink 2 liters of water",
-    stepName: "Drink 2 liters of water",
-    category: "Fitness",
-    description: "Drink 2 liters of water every day, for 30 days",
-    durationDays: 30,
-    
-  },
-  {
-    name: "Go for a walk",
-    stepName: "Go for a walk",
-    category: "Fitness",
-    description: "Go for a walk every day, for 30 days",
-    durationDays: 30,
-    
-  },
-  {
-    name: "Go to the gym",
-    stepName: "Go to the gym",
-    category: "Fitness",
-    description: "Go to the gym every day, for 30 days",
-    durationDays: 30,
-    
-  },
-  {
-    name: "Write a blog post",
-    stepName: "Write a blog post",
-    category: "Career",
-    description: "Write a blog post every day, for 30 days",
-    durationDays: 30,
-    
-  },
-  {
-    name: "Take a class",
-    stepName: "Take a class",
-    category: "Education",
-    description: "Take a class every day, for 30 days",
-    durationDays: 30,
-    
-  },
-  {
-    name: "Write an email to a friend",
-    stepName: "Write an email to a friend",
-    category: "Relationships",
-    description: "Write an email to a friend every day, for 30 days",
-    durationDays: 30,
-    
-  },
-  {
-    name: "Call a friend",
-    stepName: "Call a friend",
-    category: "Relationships",
-    description: "Call a friend every day, for 30 days",
-    durationDays: 30,
-    
-  },
-  {
-    name: "Cook a new recipe",
-    stepName: "Cook a new recipe",
-    category: "Fitness",
-    description: "Cook a new recipe every day, for 7 days",
-    durationDays: 7,
-    
-  },
-  {
-    name: "Learn a new language",
-    stepName: "Learn a new language",
-    category: "Education",
-    description: "Learn a new language every day, for 100 days",
-    durationDays: 100,
-    
-  },
-  {
-    name: "Learn a new skill",
-    stepName: "Learn a new skill",
-    category: "Education",
-    description: "Learn a new skill every day, for 7 days",
-    durationDays: 7,
-    
-  },
-  {
-    name: "Write a book",
-    stepName: "Work on my book",
-    category: "Education",
-    description: "Work on a book every day, for a year",
-    durationDays: 365,
-    
-  },
-  {
-    name: "Write a poem",
-    stepName: "Write a poem",
-    category: "Education",
-    description: "Write a poem every day, for 100 days",
-    durationDays: 100,
-    
-  },
-  {
-    name: "Learn to play an instrument",
-    stepName: "Learn to play an instrument",
-    category: "Hobbies",
-    description: "Learn to play an instrument every day, for 100 days",
-    durationDays: 100,
-    
-  },
-  {
-    name: "Learn to dance",
-    stepName: "Learn to dance",
-    category: "Hobbies",
-    description: "Learn to dance every day, for 10 days",
-    durationDays: 10,
-    
-  },
-  {
-    name: "Visit a place",
-    stepName: "Visit a place",
-    category: "Hobbies",
-    description: "Visit a place every day, for 7 days",
-    durationDays: 7,
-    
-  },
-  {
-    name: "Limit your screen time to 1.5 hours",
-    stepName: "Limit your screen time to 1.5 hours",
-    category: "Mental health",
-    description: "Limit your screen time to 1.5 hours every day, for 30 days",
-    durationDays: 30,
-    
-  },
-
-  {
-    name: "Create a retirement plan",
-    stepName: "Create a retirement plan",
-    category: "Finance",
-    description: "Research and create a retirement plan within the next month",
-    durationDays: 30,
-    
-  },
-
-  {
-    name: "Pay off credit card debt",
-    stepName: "Pay off credit card debt",
-    category: "Finance",
-    description: "Pay off all credit card debt within 3 months",
-    durationDays: 90,
-    
-  },
-  {
-    name: "Save $1000",
-    stepName: "Save $1000",
-    category: "Finance",
-    description: "Save $1000 in your savings account within 3 months",
-    durationDays: 90,
-    
-  },
-];
-
-export const routines = [
-  {
-    name: "Gain muscle (without equipment)",
-    emoji:
-      "https://cdn.jsdelivr.net/npm/emoji-datasource-apple/img/apple/64/1f4aa.png",
-    timeOfDay: 14,
-    daysOfWeek: "[true, true, true, true, true, true, true]",
-    durationDays: 100,
-    category: "Fitness",
-    items: [
-      {
-        name: "10 push-ups every day",
-        stepName: "Do 10 push-ups today",
-        category: "Fitness",
-        description: "Do 10 push-ups every day",
-        durationDays: 100,
-      },
-      {
-        name: "50 squats every day",
-        stepName: "Do 50 squats today",
-        category: "Fitness",
-        description: "Perform 50 squats every day to build leg muscles.",
-        durationDays: 30,
-      },
-      {
-        name: "50 crunches every day",
-        stepName: "Do 50 crunches today",
-        category: "Fitness",
-        description: "Perform 50 crunches every day to build core muscles.",
-        durationDays: 60,
-      },
-    ],
-  },
-];
