--- pages/api/user/checkIns/index.ts
@@ -1,15 +1,0 @@
-import { prisma } from "@/lib/server/prisma";
-
-// Checks if the user has already done today's daily check-in
-export default async function handler(req, res) {
-  const data = await prisma.dailyCheckIn.findMany({
-    where: {
-      AND: [
-        { date: new Date(req.query.date) },
-        { userId: req.query.userIdentifier },
-      ],
-    },
-  });
-
-  res.json(data);
-}
