--- pages/api/user/coach/goals/delete.ts
@@ -1,9 +1,0 @@
-import { prisma } from "@/lib/server/prisma";
-
-export default async function handler(req, res) {
-  const data = await prisma.routineItem.delete({
-    where: { id: req.query.id },
-  });
-
-  res.json(data);
-}
