+++ push-server/src/dailyNudge.ts
@@ -23,9 +23,6 @@ const handler = {
       params
     );
 
-    // Daily check in nudge
-    await fetch("https://my.dysperse.com/api/cron/dailyCheckInNudge", params);
-
     // Task thingy
     await fetch("https://my.dysperse.com/api/cron/taskReminder", params);
