--- styles/coach.scss
@@ -1,56 +1,0 @@
-.animate-trophy {
-  animation: trophy 2s ease-in-out forwards;
-  transform: scale(0) rotate(-45deg);
-}
-
-@keyframes trophy {
-  0% {
-    transform: scale(0) rotate(-45deg);
-  }
-  50% {
-    transform: scale(1.1) rotate(0deg);
-  }
-  75% {
-    transform: scale(1) rotate(0deg);
-  }
-  100% {
-    transform: scale(1) rotate(0deg) translateY(-50px);
-  }
-}
-
-.slide-in-bottom {
-  bottom: -500px;
-  animation: slide-in-bottom 2.5s cubic-bezier(0.25, 0.46, 0.45, 0.94) both;
-}
-
-@keyframes slide-in-bottom {
-  0% {
-    bottom: -500px;
-  }
-  50% {
-    bottom: -500px;
-  }
-  100% {
-    bottom: 0;
-  }
-}
-
-@keyframes jiggle {
-  0% {
-    transform: rotate(-0.5deg);
-  }
-  50% {
-    transform: rotate(0.5deg);
-  }
-}
-
-@media (min-width: 600px) {
-  @keyframes jiggle {
-    0% {
-      transform: rotate(-0.1deg);
-    }
-    50% {
-      transform: rotate(0.1deg);
-    }
-  }
-}
